Fixing bug with aliases in initial build-out
The CFN event rules were failing because they need the lambda alias to
exist. Solution was to define the aliases in the template pointed to
"$LATEST" version, then do an immideate `release` after stack creation.
This simplified things a bit because the `stack.create` task no longer
has to do the intitial alias creation bit.